+++ lib/ui/util/formatters.dart
@@ -507,7 +507,8 @@ String getThemeCurrencyMode(BuildContext context) {
 // card time format:
 // String CARD_TIME_FORMAT = "MMM d, h:mm a";
 // ignore: non_constant_identifier_names
-String CARD_TIME_FORMAT = "MMM dd, HH:mm";
+// String CARD_TIME_FORMAT = "MMM dd, HH:mm";
+String CARD_TIME_FORMAT = "y/M/d, HH:mm";
 String CARD_HOUR_TIME_FORMAT = "HH:mm";
 String getTimeAgoString(BuildContext context, int epochTime) {
   String timeStr = DateFormat(CARD_TIME_FORMAT).format(DateTime.fromMillisecondsSinceEpoch(epochTime * 1000));