+++ lib/ui/util/ui_util.dart
@@ -517,11 +517,12 @@ class UIUtil {
       Align(
         alignment: Alignment.topCenter,
         child: Container(
-          margin: EdgeInsets.symmetric(vertical: MediaQuery.of(context).size.height * 0.07, horizontal: 14),
+          // margin: EdgeInsets.symmetric(vertical: MediaQuery.of(context).size.height * 0.07, horizontal: 14),
+          margin: EdgeInsets.only(top: MediaQuery.of(context).size.height * 0.85, left: 14, right: 14),
           padding: const EdgeInsets.symmetric(vertical: 15, horizontal: 15),
           width: MediaQuery.of(context).size.width - 30,
           decoration: BoxDecoration(
-            color: StateContainer.of(context).curTheme.primary,
+            color: StateContainer.of(context).curTheme.warning,
             borderRadius: BorderRadius.circular(10),
             boxShadow: [
               BoxShadow(color: StateContainer.of(context).curTheme.barrier!, offset: const Offset(0, 15), blurRadius: 30, spreadRadius: -5),